\encoding{UTF-8}
\name{llradat3}
\alias{llradat3}
\docType{data}
\title{An Artificial LLRA Data Set}
\description{
Artificial data set of 3 items, 2 time points and 2 groups for LLRA. It
is example 6 from Hatzinger and Rusch (2009).
}
\usage{llradat3}
\format{
  A data frame with 60 observations of 6 variables.
  \describe{
    \item{\code{V1}}{Answers to item 1 at time point 1}
    \item{\code{V2}}{Answers to item 2 at time point 1}
    \item{\code{V3}}{Answers to item 3 at time point 1}
    \item{\code{V4}}{Answers to item 1 at time point 2}
    \item{\code{V5}}{Answers to item 2 at time point 2}
    \item{\code{V6}}{Answers to item 3 at time point 2}
  }
}
\details{
This is a data set as described in Hatzinger and Rusch (2009).
}
\references{
Hatzinger, R. and Rusch, T. (2009) IRT models with relaxed assumptions
in eRm: A manual-like instruction. \emph{Psychology Science Quarterly}, \bold{51},
pp. 87--120.
}
\examples{
llradat3
}
\keyword{datasets}
