\name{X_rasch}
\alias{X_rasch}
\alias{X_lltm}
\alias{X_pcm}
\alias{X_lpcm}
\alias{X_rsm}
\alias{X_lrsm}
\alias{W_lpcm}

\docType{data}
\title{Data for computing extended Rasch models}
\description{Artificial data sets and design matrices for computing extended Rasch models.
}
\usage{data(X_rasch)}    %andere datensätze.
\format{Numeric matrices with subjects as rows and items as columns.
}
\examples{
data(X_rasch)
data(X_lltm)
data(X_pcm)
data(X_lpcm)
data(X_rsm)
data(X_lrsm)
data(W_lpcm)
}
\keyword{datasets}
