\name{raschdat1}
\alias{raschdat1}
\alias{raschdat2}
\alias{lltmdat1}
\alias{lltmdat2}
\alias{pcmdat}
\alias{lpcmdat}
\alias{rsmdat}
\alias{lrsmdat}

\docType{data}
\title{Data for Computing Extended Rasch Models}
\description{Artificial data sets for computing extended Rasch models.
}
\usage{data(raschdat1)}    
\format{Numeric matrices with subjects as rows, items as columns, missing values as \code{NA}.
}
\examples{
data(raschdat1)
data(raschdat2)
data(lltmdat1)
data(lltmdat2)
data(pcmdat)
data(lpcmdat)
data(rsmdat)
data(lrsmdat)
}
\keyword{datasets}
