\name{raschdat}
\alias{raschdat}
\alias{lltmdat}
\alias{pcmdat}
\alias{lpcmdat}
\alias{rsmdat}
\alias{lrsmdat}
\alias{lpcmdat}

\docType{data}
\title{Data for computing extended Rasch models}
\description{Artificial data sets for computing extended Rasch models.
}
\usage{data(raschdat)}    
\format{Numeric matrices with subjects as rows and items as columns.
}
\examples{
data(raschdat)
data(lltmdat)
data(pcmdat)
data(lpcmdat)
data(rsmdat)
data(lrsmdat)
data(lpcmdat)
}
\keyword{datasets}
